removing more shit
