+++ src/module/item/item-sheet.js
@@ -1,56 +1,9 @@
 import { PMTTRPGUtility } from "../utility.js";
-import { buildEffectSummaryGroups } from "../effects/effect-summary.js";
+import { buildEffectSummaryGroups, computeEffectSummary } from "../effects/effect-summary.js";
 
 const { ItemSheetV2 } = foundry.applications.sheets;
 const { HandlebarsApplicationMixin } = foundry.applications.api;
 const { TextEditor, FormDataExtended } = foundry.applications.ux;
-
-function computeEffectSummary(entries = [], epMax = 0) {
-  let positiveSpent = 0;
-  let negativeSpent = 0;
-  const signatureCounts = new Map();
-
-  for (const entry of entries ?? []) {
-    const cost = Math.abs(Number(entry?.cost ?? 0));
-    const stack = Math.max(1, Number(entry?.stack ?? entry?.count ?? 1));
-    const signature = [
-      entry?.effectUuid ?? '',
-      entry?.procOn ?? '',
-      entry?.procResult ?? '',
-      entry?.procStat ?? '',
-      entry?.procDice ?? '',
-      entry?.procAction ?? '',
-      entry?.procCondition ?? '',
-      entry?.mode ?? ''
-    ].join('|').toLowerCase();
-
-    signatureCounts.set(signature, (signatureCounts.get(signature) ?? 0) + 1);
-
-    if (entry?.mode === 'negative') {
-      negativeSpent += cost * stack;
-    }
-    else {
-      positiveSpent += cost * stack;
-    }
-  }
-
-  const cap = Number.isFinite(Number(epMax)) ? Number(epMax) : 0;
-  const remaining = (cap + negativeSpent) - positiveSpent;
-  const overPositive = positiveSpent > (cap + negativeSpent);
-  const overNegative = negativeSpent > cap;
-  const hasDuplicates = Array.from(signatureCounts.values()).some(count => count > 1);
-
-  return {
-    epMax: cap,
-    positiveSpent,
-    negativeSpent,
-    remaining,
-    overPositive,
-    overNegative,
-    hasDuplicates,
-    hasWarnings: overPositive || overNegative || hasDuplicates
-  };
-}
 
 /**
  * Extend the basic ItemSheet with some very simple modifications
