+++ src/templates/chat/chat-move.html
@@ -51,10 +51,5 @@
         {{> "systems/projectmoonttrpg/templates/parts/chat-buttons.html" }}
       {{/if}}
     {{/if}}
-    {{!-- Tags --}}
-    {{#if tags}}
-      <hr>
-      {{{dwTags tags}}}
-    {{/if}}
   </div>
 </section>
